+++ db/migrate/20140921012247_remove_outdated_tables.rb
@@ -1,11 +1,15 @@
+# frozen_string_literal: true
+
 class RemoveOutdatedTables < ActiveRecord::Migration
   def change
-    remove_column(:people, :fullname) rescue nil
-    
-    %w{ duplicates_racers engine_schema_info historical_names images news_items promoters racers standings users }.each do |name|
-      if table_exists?(name)
-        drop_table name
-      end
+    begin
+      remove_column(:people, :fullname)
+    rescue StandardError
+      nil
+    end
+
+    %w[ duplicates_racers engine_schema_info historical_names images news_items promoters racers standings users ].each do |name|
+      drop_table name if table_exists?(name)
     end
   end
 end
